+++ NEWS.md
@@ -1,3 +1,8 @@
+# hyd1d 0.5.1
+
+* adapt BfG's new corporate design
+* remove the dependency to the orphaned package plotrix
+
 # hyd1d 0.5.0
 
 * use httr2 to interact with the pegelonline rest api